Create README.md
So I have this simple calculator for Python And it starts up just fine. The only problem I have is that it doesn't want to finish the equations that I put in. Example, every time I would put 8 + 4, And press equal It doesn't do anything and I can't for the life of me figure out what's wrong with the program.